requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.1
black==19.3b0
celery==4.3.0
Click==7.0
coverage==4.5.4
environs==5.2.0
factory-boy==2.12.0
feedparser==5.2.1
flake8==3.7.8
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-isort==2.7.0
Flask==1.1.1
Flask-Caching==1.7.2
Flask-CORS==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
isort==4.3.21
lxml==4.4.1
pdbpp==0.10.0
pep8-naming==0.8.2
psycopg2-binary==2.7.4
pydocstyle==3.0.0
pyopenssl==19.0.0
pytest==5.2.1
pytest-cov==2.8.1
redis==3.3.11
spacy==2.2.1
SQLAlchemy==1.3.6
WebTest==2.0.33
Werkzeug==0.15.5
WTForms==2.2.1